Refactor HistoricalStorage into cohesive submodules
